Set `headerpad_max_install_names` in `build.zig`
This is needed in order to create a binary distribution of FastFEC. The
built binary has references to library dependencies on the build machine
that need to be rewritten to be suitable for the machine it will be
installed in. This rewriting will fail without setting
`headerpad_max_install_names`.

Zig also has this code in their `build.zig` [1].

[1] https://github.com/ziglang/zig/blob/436e99d13ba188412b8a431b69cc9ff29c6bec4a/build.zig#L551-L554