ref: Move mapping regexes to separate, static struct
It was weird to store these regexes in the persistentMemory. We
only have one version of the mappings
per compilation (and runtime) so
we can just store them statically.

Next up would be to refactor the lookupMappings()
function from fec.c
to mappings.c

While I'm at it, clean up the regex compilation to share
the helper function we already have.
This involved creating
regex.h to hold that shared regexCompile()