test: ref: Overhaul python testing
Before, the testing was a bit half-baked.
We didn't actually check the contents of the output files.
We also weren't very thorough, and only checked some APIs on
some examples. Now we test all APIs on all examples.

Some example .fec files that were tracked in git weren't actually
getting tested at all. For example,
the 13360.fec file actually produces a segfault!
We need to track down the issue and fix it.
It is skipped for now.

These changes also revealed an inconsistency in  howform types
are normalized. See `python/tests/filings