ref: explicitly pass NULL for types for header
We haven't yet set ctx->types yet, so it will always be NULL at
this point. But make this explicit.

We set ctx->types once we start parsing lines. It would be
irrelevant to even try to pass something here, because
there is no entry in the `types` lookup
table for the `hdr` form.