ref: remove unused FEC_CONTEXT.currentLineLength, LINE_INFO.length
These were mistakenly used before, when really we should have been using
ctx->persistentMemory->line->n
Now I've switched us to that everywhere, so it's safe to remove this.

This also has the cascading effect of allowing us to simplify the API
of decodeLine() so it no longer returns anything,
it just decodes the line.

This also fixes a bug in iso_8859_1_to_utf_8()
where we didn't increment length when we should have.
I'm amazed this didn't crash before?
It makes me nervous that we're not using it correctly
elsewhere.